# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MegaByte's Circuit Systems API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SubmitResponse {
  success: boolean;
  message: string;
  id?: string;
}

export interface ContactInput {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  serviceType?: string | null;
  message: string;
}

export interface QuoteInput {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  pcbType: string;
  layers: number;
  /** @nullable */
  boardWidth?: number | null;
  /** @nullable */
  boardHeight?: number | null;
  quantity: number;
  /** @nullable */
  thickness?: string | null;
  /** @nullable */
  copperWeight?: string | null;
  /** @nullable */
  surfaceFinish?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface QuoteCalculationInput {
  pcbType: string;
  layers: number;
  boardWidth: number;
  boardHeight: number;
  quantity: number;
  /** @nullable */
  thickness?: string | null;
  /** @nullable */
  copperWeight?: string | null;
  /** @nullable */
  surfaceFinish?: string | null;
}

export type QuoteCalculationBreakdown = {
  baseCost: number;
  layerSurcharge: number;
  finishSurcharge: number;
};

export interface QuoteCalculation {
  estimatedCost: number;
  leadTime: string;
  currency: string;
  breakdown?: QuoteCalculationBreakdown;
}