# lib/api-zod/src/generated/types/quoteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MegaByte's Circuit Systems API
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteInput {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  pcbType: string;
  layers: number;
  /** @nullable */
  boardWidth?: number | null;
  /** @nullable */
  boardHeight?: number | null;
  quantity: number;
  /** @nullable */
  thickness?: string | null;
  /** @nullable */
  copperWeight?: string | null;
  /** @nullable */
  surfaceFinish?: string | null;
  /** @nullable */
  notes?: string | null;
}